refactor: Microservices structure
