+++ docs/insomnia_2024-01-19.json
@@ -1,0 +1,1 @@
+{"_type":"export","__export_format":4,"__export_date":"2024-01-19T13:02:51.145Z","__export_source":"insomnia.desktop.app:v8.5.1","resources":[{"_id":"req_1ffff52fe7a94e6bbe6ed1ebba587b48","parentId":"fld_85b3cd05c877442b90357407d6c2566b","modified":1705596935142,"created":1705591326847,"url":"http://localhost:8080/persons","name":"save","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"name\": \"Isabella Nunes\",\n\t\"email\": \"isadfrn1@gmail.com\",\n\t\"cpf\": \"098789789141\",\n\t\"phone\": \"065787898765\",\n\t\"role\": \"ADMIN\",\n\t\"password\": \"1234567890\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/8.5.1"}],"authentication":{},"metaSortKey":-1705440094430,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"fld_85b3cd05c877442b90357407d6c2566b","parentId":"wrk_7d5eec3bb59247c2a42a2870434fb9c7","modified":1705590204268,"created":1705432043318,"name":"persons","description":"","environment":{},"environmentPropertyOrder":null,"metaSortKey":-1705525228183,"_type":"request_group"},{"_id":"wrk_7d5eec3bb59247c2a42a2870434fb9c7","parentId":null,"modified":1705431983172,"created":1705431983172,"name":"carinspection","description":"","scope":"collection","_type":"workspace"},{"_id":"req_a75bcda5c7e8454fbb6ff0e61630030b","parentId":"fld_85b3cd05c877442b90357407d6c2566b","modified":1705532071215,"created":1705440094330,"url":"http://localhost:8080/persons","name":"list","description":"","method":"GET","body":{},"parameters":[],"headers":[{"name":"User-Agent","value":"insomnia/8.5.1"}],"authentication":{},"metaSortKey":-1705440094330,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_9de38b0218c142338742cf25a356192e","parentId":"fld_c82ae9a1dcaa4371bf2146fe7e836e6f","modified":1705596972560,"created":1705525249057,"url":"http://localhost:3333/portfolios","name":"save","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"userId\": \"1\",\n  \"vehicleType\": \"CAR\",\n  \"brand\": \"Fiat\",\n  \"model\": \"Uno\",\n  \"licensePlate\": \"GOT1234\",\n  \"chassis\": \"1HGCM82633A004352\",\n  \"manufactureYear\": \"2000\",\n\t\"vehicleImages\": [\"https://2.bp.blogspot.com/-o-0_K2bkKlA/TocnWx9M6NI/AAAAAAAAAQ8/UblhrQv5dYQ/w1200-h630-p-k-no-nu/uno-mille-way-vermelho.jpg\"]\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/8.5.1"}],"authentication":{},"metaSortKey":-1705541356984,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"fld_c82ae9a1dcaa4371bf2146fe7e836e6f","parentId":"wrk_7d5eec3bb59247c2a42a2870434fb9c7","modified":1705531926051,"created":1705525228083,"name":"portfolios","description":"","environment":{},"environmentPropertyOrder":null,"metaSortKey":-1705525228083,"_type":"request_group"},{"_id":"req_6dffe13e87be41e89103373b74e9e8f6","parentId":"fld_c82ae9a1dcaa4371bf2146fe7e836e6f","modified":1705541703853,"created":1705541356884,"url":"http://localhost:3333/portfolios","name":"list","description":"","method":"GET","body":{},"parameters":[],"headers":[{"name":"User-Agent","value":"insomnia/8.5.1"}],"authentication":{},"metaSortKey":-1705541356884,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"env_de06cadad068ef391b7732f2755a13c617420581","parentId":"wrk_7d5eec3bb59247c2a42a2870434fb9c7","modified":1705432021091,"created":1705431983173,"name":"Base Environment","data":{"BASE_URL":"http://localhost:8080"},"dataPropertyOrder":{"&":["BASE_URL"]},"color":null,"isPrivate":false,"metaSortKey":1705431983173,"_type":"environment"},{"_id":"jar_de06cadad068ef391b7732f2755a13c617420581","parentId":"wrk_7d5eec3bb59247c2a42a2870434fb9c7","modified":1705431983174,"created":1705431983174,"name":"Default Jar","cookies":[],"_type":"cookie_jar"}]}
