--- java-server/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.hibernate.show-sql=true
-spring.datasource.url=jdbc:postgresql://localhost:5432/carinspect
-spring.datasource.username=postgres
-spring.datasource.password=root
-springdoc.api-docs.path=/api-docs
