--- node-server/docker-compose.yml
@@ -1,20 +1,0 @@
-version: "3.8"
-services:
-  node-server:
-    build: .
-    ports:
-      - "8080:8080"
-    depends_on:
-      - db
-  db:
-    image: mongo
-    ports:
-      - "27017:27017"
-    environment:
-      MONGO_INITDB_ROOT_USERNAME: root
-      MONGO_INITDB_ROOT_PASSWORD: root
-      MONGO_INITDB_DATABASE: carinspect
-    volumes:
-      - data:/var/lib/mongo/data
-volumes:
-  data:
