+++ persons/src/main/java/com/isadfrn/carinspection/entities/dto/PersonDTO.java
@@ -1,0 +1,29 @@
+package com.isadfrn.carinspection.entities.dto;
+
+import com.isadfrn.carinspection.entities.enums.Role;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PersonDTO {
+    private Long id;
+    @NotNull(message = "Name is required")
+    private String name;
+    @NotNull(message = "Email is required")
+    @Email(message = "Email should be valid")
+    private String email;
+    @NotNull(message = "CPF is required")
+    private String cpf;
+    @NotNull(message = "Phone is required")
+    private String phone;
+    private Role role = Role.USER;
+    @NotNull(message = "Password is required")
+    private String password;
+}
