+++ persons/src/main/java/com/isadfrn/carinspection/entities/enums/Role.java
