+++ persons/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.show-sql=true
+spring.datasource.url=jdbc:postgresql://localhost:5432/carinspect
+spring.datasource.username=root
+spring.datasource.password=root
+springdoc.api-docs.path=/api-docs
